% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_contrast_rm.R
\name{t_contrast_rm}
\alias{t_contrast_rm}
\title{t-test contrasts for one-way repeated measures}
\usage{
t_contrast_rm(df, subject, dv, condition, A_levels, B_levels,
  contrast_weights)
}
\arguments{
\item{df}{A long format dataframe}

\item{subject}{name of the subject column}

\item{dv}{name of the dependent variable column}

\item{condition}{name of the condition for the one-way manipulation}

\item{A_levels}{character vector, including the names of the levels for the first side of the contrast}

\item{B_levels}{character vector, including the names of the levels for the second side of the contrast}

\item{contrast_weights}{numeric vector, including the contrast weights for all levels, in the same order as they are listed in A and B levels}
}
\value{
t.test summary
}
\description{
t-test contrasts for one-way repeated measures
}
